+++ src/Phorkus.Consensus/CommonCoin/CoinId.cs
@@ -6,16 +6,16 @@ namespace Phorkus.Consensus.CommonCoin
 {
     public class CoinId : IProtocolIdentifier
     {
-        public CoinId(uint era, uint agreement, uint epoch)
+        public CoinId(ulong era, ulong agreement, ulong epoch)
         {
             Era = era;
             Agreement = agreement;
             Epoch = epoch;
         }
 
-        public uint Era { get; }
-        public uint Agreement { get; }
-        public uint Epoch { get; }
+        public ulong Era { get; }
+        public ulong Agreement { get; }
+        public ulong Epoch { get; }
 
         public IEnumerable<byte> ToByteArray()
         {
